% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_data_boot_one.R
\name{create_data_boot_one}
\alias{create_data_boot_one}
\title{create one bootsraping data}
\usage{
create_data_boot_one(boot_count, Indata)
}
\arguments{
\item{boot_count}{the count number of bootstrapping}

\item{Indata}{the data that you used}
}
\description{
This function create one data for bootsrapping
}
