% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_vcovmatrix.R
\name{create_vcovmatrix}
\alias{create_vcovmatrix}
\title{create_vcovmatrix}
\usage{
create_vcovmatrix(out.reg, s.reg, q.reg)
}
\arguments{
\item{out.reg}{the output of the R function 'clm' i.e. the regression output of the outcome}

\item{s.reg}{the output of the R functiom 'lm' under the model of 'S' (the 2nd medistor)}

\item{q.red}{the output of the R function 'lm' under the model of 'Q' (the 1st medistor)}
}
\description{
create the variance-covariance matrix of theta
}
