% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ord_mediation_analysis_pal_sc.R
\name{Ord_mediation_analysis_pal_sc}
\alias{Ord_mediation_analysis_pal_sc}
\title{Ord_mediation_analysis for sensitivity analysis}
\usage{
Ord_mediation_analysis_pal_sc(
  Indata,
  n_cate,
  confounders,
  intv = 4,
  intval = c(0, 1),
  nb = 0,
  w_c,
  n.core = 1
)
}
\arguments{
\item{Indata:}{Input data with outcome, exposure, mediators, and covariates as columns. The number of rows equals to the number of people in the data you used.}

\item{n_cate:}{the number of categories of the outcome}

\item{confounders:}{the values of confounder}

\item{intv:}{Number of intervention, only 4.}

\item{intval:}{the value of exposure used in Intervention , Default is c(0,1).}

\item{nb:}{Number of bootstrapping. Default is 0 (no bootstrapping applied).}

\item{n_core:}{number of cores that will be used in parallel computing. default=1}

\item{w_c<-}{cate of Y that user have instercting in. If all in let wc="all",wc will set to be seq(1,n.cate,by=1).}
}
\value{
please see the help docunment of function "Ord_mediation_analysis_pal"
}
\description{
Apply mediation analysis for non-rare orinal/multinomial outcome with two gaussian mediators
for plot of bias
}
