% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PSEa_wiboot_c.R
\name{PSEa_wiboot_c}
\alias{PSEa_wiboot_c}
\title{PSE with bootstrap}
\usage{
PSEa_wiboot_c(
  data_boot,
  Para,
  n.para,
  n.cate,
  ITE,
  theta.hat,
  sig.hat,
  Vcov.matrix,
  intv,
  intval,
  confounders,
  a,
  nb,
  n.core
)
}
\arguments{
\item{Para}{the parameter that will trated as nominal variables. Seeting as "~W+Q+S". The detail please see the R function 'clm'}

\item{ITE}{the total effect}

\item{theta.hat}{the estimate of theta (all of the parameters of the models)}

\item{sig.hat}{the eastimate of sigma (all of the variance of the parameters)}

\item{Vcov.matrix}{the variance-covariance matrix of theta}

\item{intv}{number of interventions. Default is 4.}

\item{intval}{the values of exposure used for intervention. Default is c(0,1)}

\item{confounders}{the values of confounders, def:c()}

\item{a}{the value of outcome}

\item{data.boot}{the data set}

\item{n_para}{the number of pararmeters that used in the model of outcome}

\item{n.cata}{the number of categories of the outcome}

\item{nb:}{number of bootstrapping}

\item{n.core:}{number of cores that will be used}
}
\description{
PSE with bootstrap
}
