% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_ord.R
\name{rd_ord}
\alias{rd_ord}
\title{this function calculate the PSEs under risk difference scale and 95\% confidence interval}
\usage{
rd_ord(pa, pb, V)
}
\arguments{
\item{pa}{rho under one intervention}

\item{pb}{rho under another intervention}

\item{V}{the variance-covariance matrix of theta}
}
\description{
this function calculate the PSEs under risk difference scale and 95\% confidence interval
}
