% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rho.R
\name{rho}
\alias{rho}
\title{rho}
\usage{
rho(TH, SH, n.para, Name = c("W", "Q", "S"), n.cate1, a, w, confounders = NULL)
}
\arguments{
\item{TH}{theta; all of the parameters of the model}

\item{SH}{sigma: all of the estimates of the variance of the mediators}

\item{n.para}{number of parameters}

\item{Name}{names of exposure and mediators, def:c("W","Q","S")}

\item{a}{outcome}

\item{w}{the values of intervention}

\item{confounders}{the values of confounder, def:NULL}
}
\description{
The probability of the counterfactual outcome of Y
}
\details{
This function allows you to calculate the probability of the counterfactual outcome of Y.
}
\examples{
rho(theta.hat,sig.hat,n.para,Name=c("W","Q","S"),n.cate1,a,w0000,confounders)
}
\keyword{counterfactual}
