% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rr_ord.R
\name{rr_ord}
\alias{rr_ord}
\title{this function calculate the PSEs under risk ratio scale and 95\% confidence interval}
\usage{
rr_ord(pa, pb, V)
}
\arguments{
\item{pa}{rho under one intervention}

\item{pb}{rho under another intervention}

\item{V}{the variance-covariance matrix of theta}
}
\description{
this function calculate the PSEs under risk ratio scale and 95\% confidence interval
}
