% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_medi_cate_data.R
\name{sim_medi_cate_data}
\alias{sim_medi_cate_data}
\title{simulate a data}
\usage{
sim_medi_cate_data(
  sample_size,
  exp_cont,
  med_cont,
  n_cate,
  deltaq,
  alphas,
  betay,
  coa = c(),
  j
)
}
\arguments{
\item{sample_size}{the number of case wanted to be create}

\item{exp_cont}{the mean and standard deviations of exposure; only consider continuous exposure}

\item{med_cont}{vector of standard deviations of the mediators; only consider continous mediators}

\item{n_cate}{the number of categories of the outcome}

\item{deltaq}{the true value of the parameter of the model of Q}

\item{alphas}{the true value of the parameter of the model of S}

\item{betay}{the true value of the parameter of the model of Y}

\item{coa}{the true value of mean and variance of the covariates seeting as c(mean(coa1),sd(coa1),mean(coa2),sd(coa2),...); default:coa=C()}

\item{j}{is the number to set seed; set.seed(j)}
}
\value{
Y the outcome

W the exposure

Q the 1st mediator

S the 2nd mediator

Coa1 the 1st covariate
}
\description{
this function to generate simulation data
}
\details{
how the mediators and exposures are generated
W=exposure, Q=mediator one right after W, S=second mediator closest to outcome, Y=outcome
}
\examples{
sample_size<- 500
exp_cont<- c(0,1)
med_cont<- c(0.23,0.23)
n_cate<- 3
deltaq<- c(0.1,0.5) #Q c(d0,dw)
alphas<- c(0.2,0.5,0.5) #S c(al0,alw,alq)
betay1<- c(-1.8, 1.0, 1.2, 1.3) #c(b01,bw1,bq1,bs1)
betay2<- c(0.8, 0.8, 1.1, 1.2) #c(b02,bw2,bq2,bs2)
betay<- rbind(betay1,betay2)
coa<- c()
j<-56 #sample(1,1:1e+09)
sim_cover<-lapply(j,"sim_medi_cate_data",sample_size=sample_size,exp_cont=exp_cont,med_cont=med_cont,n_cate=n_cate,deltaq=deltaq,alphas=alphas,betay=betay,coa=coa)
}
