% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/total_a_r.R
\name{total_a_r}
\alias{total_a_r}
\title{total_a_r}
\usage{
total_a_r(TEa, TEa1, intval = c(0, 1), confounders)
}
\arguments{
\item{TEa}{total effect under outcome=a}

\item{TEa1}{total effect under outcome=a-1}

\item{intval}{the values of exposure used for intervention. Default is c(0,1).}

\item{confounders}{values of Confounders}
}
\description{
total_a_r the risk of total effet undera the value of outcome is a
}
