% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Ord_mediation_analysis_pal.R
\name{Ord_mediation_analysis_pal}
\alias{Ord_mediation_analysis_pal}
\title{Ord_mediation_analysis
multinomial outcome mediation analysis with two ordered mediators}
\usage{
Ord_mediation_analysis_pal(
  Indata,
  n_cate,
  confounders,
  intv = 4,
  intval = c(0, 1),
  nb = 0,
  n_core
)
}
\arguments{
\item{Indata:}{Input data with outcome, exposure, mediators, and covariates as columns. The number of rows equals to the number of people in the data you used.}

\item{n_cate:}{the number of categories of the outcome}

\item{confounders:}{the values of confounder}

\item{intv:}{number of interventions. Default is 4 and only 4.}

\item{intval:}{the values of exposure used for intervention. Default is c(0,1).}

\item{nb:}{Number of bootstrapping. Default is 0 (i.e. no bootstrapping applied).}

\item{n_core:}{number of cores that will be used}
}
\description{
Apply mediation analysis for non-rare orinal/multinomial outcome with two gaussian mediators
}
